Split used models for readability
